% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layer_factory.R
\name{add_layer_factory}
\alias{add_layer_factory}
\title{Function Factory of the \code{add_layer} Utility}
\usage{
add_layer_factory()
}
\value{
a function that builds up the static ggplot to be used by gganimate
}
\description{
Function Factory of the \code{add_layer} Utility
}
\examples{
add_layer <- add_layer_factory()
}
