% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_2d_alt.R
\name{pois_2d_alt}
\alias{pois_2d_alt}
\title{Simulating a 2D Poisson Process}
\usage{
pois_2d_alt(b, h, lambda)
}
\arguments{
\item{b}{base of the rectangle}

\item{h}{height of the rectangle}

\item{lambda}{Poisson rate}
}
\value{
x-coordinates of the Poisson points
}
\description{
Workhorse function for simulating a 2D Poisson process
over a rectangular region with its lower-left corner at the origin.
This implementation assigns the x-coordinates from a uniform distribution.
}
\examples{
pois_2d_alt(10, 100, 0.01)
}
