% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_2d_simp.R
\name{pois_2d_simp}
\alias{pois_2d_simp}
\title{Simulating a 2D Poisson Process}
\usage{
pois_2d_simp(b, h, lambda)
}
\arguments{
\item{b}{base of the rectangle}

\item{h}{height of the rectangle}

\item{lambda}{Poisson rate}
}
\value{
x-coordinates of the Poisson points
}
\description{
Workhorse function for simulating a 2D Poisson process
over a rectangular region with its lower-left corner at the origin.
This implementation determines the total number of points first and
then assign coordinates uniformly.
}
\examples{
pois_2d_simp(10, 100, 0.01)
}
