% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_2d_verifier.R
\name{pois_2d_verifier}
\alias{pois_2d_verifier}
\title{Verifier for the 2D Poisson Process Simulations}
\usage{
pois_2d_verifier(n_runs, b, h, lambda, n_parts, pois_2d_fn)
}
\arguments{
\item{n_runs}{number of runs of the 2D Poisson process}

\item{b}{base of the rectangle}

\item{h}{height of the rectangle}

\item{lambda}{Poisson rate}

\item{n_parts}{Number of parts into which the region is partitioned}

\item{pois_2d_fn}{The implementation of the workhorse Poisson simulator to be used}
}
\value{
n_runs of the x-coordinates of the 2D Poisson process
}
\description{
Verifier for the 2D Poisson Process Simulations
}
\examples{
pois_2d_verifier(1000, 10, 100, 0.01, 5, pois_2d_simp)
}
