% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_counter.R
\name{pois_counter}
\alias{pois_counter}
\title{Utility Function for Counting the Poisson points}
\usage{
pois_counter(interval, pois_runs)
}
\arguments{
\item{interval}{an interval}

\item{pois_runs}{n_runs of x-coordinates of the 2D Poisson process}
}
\value{
number of Poisson points from each run that fall within the interval
}
\description{
Utility Function for Counting the Poisson points
}
